backend/requirements.txt
streamlit
fastapi
uvicorn
requests
cloudscraper
beautifulsoup4
selenium
webdriver-manager
google-generativeai
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
oauth2client
reportlab
python-dotenv
schedule
duckduckgo-search